# ref: implemented bit encoded typed index