… core/include/detray/core/detail/tuple_container.hpp