Implements the typed index the same way as the barcode and thus saves a tiny bit of memory in the surface.