@niermann999 2023-04-26
Current main: 
```
2023-04-26T17:49:18+02:00
Running ./bin/detray_array_propagator_cuda
Run on (48 X 1794.75 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x24)
  L1 Instruction 32 KiB (x24)
  L2 Unified 512 KiB (x24)
  L3 Unified 32768 KiB (x4)
Load Average: 2.06, 4.93, 4.62
--------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------
CPU unsync propagation/8        171211 ns       170970 ns         4178 TracksPropagated=374.334k/s
CPU unsync propagation/16       521768 ns       520423 ns         1306 TracksPropagated=491.908k/s
CPU unsync propagation/32      1956104 ns      1949970 ns          358 TracksPropagated=525.136k/s
CPU unsync propagation/64      7798015 ns      7773820 ns           90 TracksPropagated=526.897k/s
CPU unsync propagation/128    32019734 ns     19671379 ns           36 TracksPropagated=832.885k/s
CPU unsync propagation/256   118890206 ns     69736511 ns           10 TracksPropagated=939.766k/s
CPU sync propagation/8          172670 ns       172323 ns         4069 TracksPropagated=371.395k/s
CPU sync propagation/16         519903 ns       518584 ns         1361 TracksPropagated=493.652k/s
CPU sync propagation/32        1983447 ns      1970763 ns          360 TracksPropagated=519.596k/s
CPU sync propagation/64        7487306 ns      7465617 ns           94 TracksPropagated=548.649k/s
CPU sync propagation/128      29933224 ns     19279798 ns           36 TracksPropagated=849.801k/s
CPU sync propagation/256     117680525 ns     67486125 ns           10 TracksPropagated=971.103k/s
CUDA unsync propagation/8      6019979 ns      6002665 ns          105 TracksPropagated=10.6619k/s
CUDA unsync propagation/16     5488754 ns      5470323 ns          127 TracksPropagated=46.798k/s
CUDA unsync propagation/32     6110564 ns      6088247 ns          114 TracksPropagated=168.193k/s
CUDA unsync propagation/64     9029445 ns      9003889 ns           77 TracksPropagated=454.915k/s
CUDA unsync propagation/128   10472803 ns     10435197 ns           67 TracksPropagated=1.57007M/s
CUDA unsync propagation/256   19765203 ns     19568980 ns           35 TracksPropagated=3.34897M/s
CUDA sync propagation/8        6105964 ns      6089671 ns          111 TracksPropagated=10.5096k/s
CUDA sync propagation/16       5706486 ns      5687105 ns          122 TracksPropagated=45.0141k/s
CUDA sync propagation/32       5482416 ns      5462870 ns          127 TracksPropagated=187.447k/s
CUDA sync propagation/64       7935336 ns      7909590 ns           88 TracksPropagated=517.852k/s
CUDA sync propagation/128      9509503 ns      9478637 ns           74 TracksPropagated=1.72852M/s
CUDA sync propagation/256     17817823 ns     17625174 ns           39 TracksPropagated=3.71832M/s
```


This PR:
```
2023-04-26T17:58:06+02:00
Running ./bin/detray_array_propagator_cuda
Run on (48 X 1796.6 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x24)
  L1 Instruction 32 KiB (x24)
  L2 Unified 512 KiB (x24)
  L3 Unified 32768 KiB (x4)
Load Average: 4.59, 5.94, 5.29
--------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------
CPU unsync propagation/8        160524 ns       160331 ns         4442 TracksPropagated=399.175k/s
CPU unsync propagation/16       539175 ns       536062 ns         1000 TracksPropagated=477.557k/s
CPU unsync propagation/32      1808852 ns      1803667 ns          392 TracksPropagated=567.732k/s
CPU unsync propagation/64      6938626 ns      6920180 ns          102 TracksPropagated=591.892k/s
CPU unsync propagation/128    27348719 ns     18738615 ns           37 TracksPropagated=874.344k/s
CPU unsync propagation/256   109809671 ns     64817145 ns           10 TracksPropagated=1011.09k/s
CPU sync propagation/8          165071 ns       164884 ns         4262 TracksPropagated=388.153k/s
CPU sync propagation/16         479790 ns       478580 ns         1466 TracksPropagated=534.916k/s
CPU sync propagation/32        1790455 ns      1785406 ns          392 TracksPropagated=573.539k/s
CPU sync propagation/64        7101765 ns      6993001 ns          101 TracksPropagated=585.728k/s
CPU sync propagation/128      27765601 ns     18774215 ns           37 TracksPropagated=872.686k/s
CPU sync propagation/256     109897593 ns     65920100 ns           10 TracksPropagated=994.173k/s
CUDA unsync propagation/8      5776547 ns      5760714 ns          111 TracksPropagated=11.1097k/s
CUDA unsync propagation/16     5173643 ns      5158419 ns          134 TracksPropagated=49.6276k/s
CUDA unsync propagation/32     6003730 ns      5985200 ns          115 TracksPropagated=171.089k/s
CUDA unsync propagation/64     8255206 ns      8231151 ns           85 TracksPropagated=497.622k/s
CUDA unsync propagation/128    9482167 ns      9450969 ns           73 TracksPropagated=1.73358M/s
CUDA unsync propagation/256   19001364 ns     18818062 ns           37 TracksPropagated=3.48261M/s
CUDA sync propagation/8        5742617 ns      5728399 ns          118 TracksPropagated=11.1724k/s
CUDA sync propagation/16       5435867 ns      5418707 ns          128 TracksPropagated=47.2437k/s
CUDA sync propagation/32       5304865 ns      5289040 ns          131 TracksPropagated=193.608k/s
CUDA sync propagation/64       7490688 ns      7467173 ns           94 TracksPropagated=548.534k/s
CUDA sync propagation/128      8868598 ns      8842740 ns           77 TracksPropagated=1.85282M/s
CUDA sync propagation/256     17187859 ns     17001785 ns           41 TracksPropagated=3.85465M/s
```