@niermann999 2023-04-26
Very minor speedup, but it is just reducing the size by 8 bytes